+++ tests/math/mathtest.cpp
@@ -59,7 +59,7 @@ static void app(int argc, char *argv[]) {
     EXPECT(o16 == 32000);
     o16 = ntoh16(hton16(buf, 33000));
     EXPECT(o16 == 33000);
-    o16 = ntoh16(hton16(buf, (unsigned) -2));
+    o16 = ntoh16(hton16(buf, (uint16_t) -2));
     EXPECT(o16 == 65534);
 
     if (int errs = logGetMsgCount(kLogTypeError)) {
